Separate arcade provider
