+++ packages/profile/src/components/context/index.tsx
@@ -1,4 +1,3 @@
-export { ArcadeContext } from "./arcade";
 export { ConnectionContext } from "./connection";
 export { DataContext } from "./data";
 export { ThemeContext } from "./theme";
