+++ packages/profile/src/hooks/arcade.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { ArcadeContext } from "../components/context/arcade";
+import { ArcadeContext } from "@/context/arcade";
 
 /**
  * Custom hook to access the Arcade context and account information.
